♻️[Refactor] Transaction 도메인 엔티티 네이밍 변경
