+++ src/main/java/com/onenth/OneNth/domain/deal/entity/DealCompletion.java
@@ -1,28 +1,26 @@
-package com.onenth.OneNth.domain.transaction.entity;
+package com.onenth.OneNth.domain.deal.entity;
 
 import com.onenth.OneNth.domain.common.BaseEntity;
-import com.onenth.OneNth.domain.transaction.entity.enums.TradeType;
+import com.onenth.OneNth.domain.deal.entity.enums.TradeType;
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class TransactionConfirmation extends BaseEntity {
+public class DealCompletion extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "transaction_id", nullable = false)
-    private Transaction transaction;
+    @JoinColumn(name = "transaction_Confirmation_id", nullable = false)
+    private DealConfirmation dealConfirmation;
 
     @Column(nullable = false)
     private LocalDateTime tradeDate;
@@ -38,7 +36,5 @@ public class TransactionConfirmation extends BaseEntity {
 
     @Column(nullable = false)
     private boolean isAccepted;
-
-    @OneToOne(mappedBy = "transactionConfirmation", cascade = CascadeType.ALL, orphanRemoval = true)
-    private TransactionCompletion transactionCompletion;
 }
+
