+++ src/main/java/com/onenth/OneNth/domain/deal/entity/enums/TradeType.java
@@ -1,0 +1,6 @@
+package com.onenth.OneNth.domain.deal.entity.enums;
+
+public enum TradeType {
+    IN_PERSON,
+    DELIVERY
+}
