+++ src/main/java/com/onenth/OneNth/domain/deal/repository/DealRepository.java
@@ -1,0 +1,7 @@
+package com.onenth.OneNth.domain.deal.repository;
+
+import com.onenth.OneNth.domain.deal.entity.Deal;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DealRepository extends JpaRepository<Deal, Long> {
+}
