--- src/main/java/com/onenth/OneNth/domain/transaction/entity/enums/TradeType.java
@@ -1,6 +1,0 @@
-package com.onenth.OneNth.domain.transaction.entity.enums;
-
-public enum TradeType {
-    IN_PERSON,
-    DELIVERY
-}
